refactor: pkg layout
