+++ cmd/orphan.go
@@ -5,12 +5,12 @@ import (
 	"os"
 	"strings"
 
-	"github.com/autobrr/tqm/client"
-	"github.com/autobrr/tqm/config"
-	"github.com/autobrr/tqm/logger"
-	paths "github.com/autobrr/tqm/pathutils"
-	"github.com/autobrr/tqm/torrentfilemap"
-	"github.com/autobrr/tqm/tracker"
+	"github.com/autobrr/tqm/pkg/client"
+	"github.com/autobrr/tqm/pkg/config"
+	"github.com/autobrr/tqm/pkg/logger"
+	"github.com/autobrr/tqm/pkg/pathutils"
+	"github.com/autobrr/tqm/pkg/torrentfilemap"
+	"github.com/autobrr/tqm/pkg/tracker"
 
 	"github.com/dustin/go-humanize"
 	"github.com/spf13/cobra"
