+++ cmd/relabel.go
@@ -4,15 +4,15 @@ import (
 	"encoding/json"
 	"time"
 
-	"github.com/autobrr/tqm/client"
-	"github.com/autobrr/tqm/config"
-	"github.com/autobrr/tqm/expression"
-	"github.com/autobrr/tqm/hardlinkfilemap"
-	"github.com/autobrr/tqm/logger"
-	"github.com/autobrr/tqm/sliceutils"
-	"github.com/autobrr/tqm/torrentfilemap"
-	"github.com/autobrr/tqm/tracker"
+	"github.com/autobrr/tqm/pkg/client"
+	"github.com/autobrr/tqm/pkg/expression"
+	"github.com/autobrr/tqm/pkg/hardlinkfilemap"
+	"github.com/autobrr/tqm/pkg/logger"
+	"github.com/autobrr/tqm/pkg/sliceutils"
+	"github.com/autobrr/tqm/pkg/torrentfilemap"
+	"github.com/autobrr/tqm/pkg/tracker"
 
+	"github.com/autobrr/tqm/pkg/config"
 	"github.com/dustin/go-humanize"
 	"github.com/spf13/cobra"
 )
