+++ pkg/client/client.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/autobrr/tqm/expression"
+	"github.com/autobrr/tqm/pkg/expression"
 )
 
 func NewClient(clientType string, clientName string, exp *expression.Expressions) (Interface, error) {
