+++ pkg/client/tagInterface.go
@@ -1,7 +1,7 @@
 package client
 
 import (
-	"github.com/autobrr/tqm/config"
+	"github.com/autobrr/tqm/pkg/config"
 )
 
 type RetagInfo struct {
