+++ pkg/hardlinkfilemap/linkinfo_darwin.go
