+++ pkg/hardlinkfilemap/linkinfo_freebsd.go
