+++ pkg/hardlinkfilemap/linkinfo_windows.go
