+++ pkg/hardlinkfilemap/struct.go
