+++ pkg/sliceutils/fastdelete.go
