+++ pkg/sliceutils/indexofstring.go
