+++ pkg/stringutils/left.go
