+++ pkg/stringutils/seconds.go
